pyproject.toml
[project]
name = "agentic-research-briefing-rag"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "streamlit>=1.32",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "python-dotenv>=1.0",
  "requests>=2.31",

  "numpy>=1.26",
  "faiss-cpu>=1.8.0",
  "rank-bm25>=0.2.2",
  "sentence-transformers>=2.6",

  "pymupdf>=1.23",
  "trafilatura>=1.8",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["core"]
